move solvers to separate namespace; rename libsolv to libmamba, legacy to classic
