+++ dev/linux/experimental-resolver.sh
@@ -10,4 +10,4 @@ conda info
 # remove the pkg cache.  We can't hardlink from here anyway.  Having it around causes log problems.
 sudo rm -rf /opt/conda/pkgs/*-*-*
 mamba --version
-CONDA_SOLVER_LOGIC=libsolv pytest -m "not integration" -k "not TestLegacySolver" -v --splits ${TEST_SPLITS} --group=${TEST_GROUP} tests/core/test_solve.py tests/test_solvers.py
+CONDA_SOLVER_LOGIC=libmamba pytest -m "not integration" -k "not TestClassicSolver" -v --splits ${TEST_SPLITS} --group=${TEST_GROUP} tests/core/test_solve.py tests/test_solvers.py
