+++ dev/linux/unit.sh
@@ -9,4 +9,4 @@ eval "$(sudo /opt/conda/bin/python -m conda init --dev bash)"
 conda info
 # remove the pkg cache.  We can't hardlink from here anyway.  Having it around causes log problems.
 sudo rm -rf /opt/conda/pkgs/*-*-*
-pytest -m "not integration" -k "not TestLibSolvSolver" -v --splits ${TEST_SPLITS} --group=${TEST_GROUP}
+pytest -m "not integration" -k "not TestLibMambaSolver" -v --splits ${TEST_SPLITS} --group=${TEST_GROUP}
