+++ dev/windows/unit.bat
@@ -2,7 +2,7 @@
 pushd %TEMP% || goto :error
 cd \conda_src || goto :error
 CALL dev-init.bat || goto :error
-CALL pytest -m "not integration" -k "not TestLibSolvSolver" -v --splits=%TEST_SPLITS% --group=%TEST_GROUP% || goto :error
+CALL pytest -m "not integration" -k "not TestLibMambaSolver" -v --splits=%TEST_SPLITS% --group=%TEST_GROUP% || goto :error
 goto :EOF
 
 :error
