+++ tests/cli/test_cli_install.py
@@ -7,6 +7,7 @@ import pytest
 from conda.models.match_spec import MatchSpec
 from conda.exceptions import UnsatisfiableError
 from conda.gateways.disk.delete import rm_rf
+from conda.base.constants import SolverLogicChoice
 from conda.base.context import context
 
 try:
@@ -25,8 +26,8 @@ class TestCliInstall(TestCase):
         rm_rf(self.prefix)
         rm_rf(self.testenv)
 
-    @pytest.mark.skipif(context.solver_logic.value == "libsolv",
-                        reason="Conflict report / analysis is done differently with libsolv.")
+    @pytest.mark.skipif(context.solver_logic == SolverLogicChoice.LIBMAMBA,
+                        reason="Conflict report / analysis is done differently with libmaba.")
     @pytest.mark.integration
     def test_find_conflicts_called_once(self):
         bad_deps = {'python': {((MatchSpec("statistics"), MatchSpec("python[version='>=2.7,<2.8.0a0']")), 'python=3')}}
